requirements.txt
attrs==23.2.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
outcome==1.3.0.post0
pillow==11.1.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
qrcode==8.0
reportlab==4.3.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
gunicorn==20.1.0
